feat(utils): add math helpers
